64bit時に警告が出ないように修正
- 警告 C4477

git-svn-id: https://svn.osdn.net/svnroot/ttssh2/trunk@7897 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ ttssh2/ttxssh/ssh.c
@@ -8676,7 +8676,7 @@ static BOOL SSH2_scp_fromremote(PTInstVar pvar, Channel_t *c, unsigned char *dat
 			HANDLE thread;
 			unsigned int tid;
 
-			sscanf_s(data, "C%o %lld %s", &permission, &size, filename, sizeof(filename));
+			sscanf_s(data, "C%o %lld %s", &permission, &size, filename, (unsigned int)sizeof(filename));
 
 			// Windowsなのでパーミッションは無視。サイズのみ記録。
 			c->scp.filetotalsize = size;
